requirements.txt
# Core ML Libraries
tensorflow>=2.10.0
keras>=2.10.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0

# Image Processing
Pillow>=8.2.0
opencv-python-headless>=4.5.3.56
imutils>=0.5.4
albumentations>=1.3.0

# UI & Visualization
streamlit>=1.10.0
matplotlib>=3.4.3
seaborn>=0.11.1

# HTML Parsing
lxml>=4.9.1

# Model Visualization (optional but recommended)
pydot>=1.4.2
graphviz>=0.20.1

# Logging & Progress
tqdm>=4.64.0

# Ensure version compatibility on macOS (especially for urllib3/libressl issues)
urllib3<2.0
